Fix head and location of script.
The head of an HTML should always contain a `<meta charset="...">`
or `<meta http-equiv="..." content="text/html; charset=..." />`,
so that one does not have to rely on HTTP charset fields.

Also, while not as nice as having the script in the head, a
script can get placed at the end of the `<body>` element. This
also relieves one from using `window.addEventListener(...)` or
`$.ready(...)`.

This change also removes the unecessary 3rd party files.