Remove duplicate code
First of all, keep your code DRY (don't repeat yourself). If you
happen to change the logic in your constructor and forget to change
the logic in `init` later (or vice versa),  you will end up with
*interesting* results.

One could call `this.init()` after its definition, which would also
work. But I wanted to demonstrate that a constructor is, after all,
just a function, which can be called with `.call`.

This enables one to reuse the original constructor completely. __BUT__
there's one draw back: the constructor has to reset all counter like
variables to a "sane" value (e.g. if you expect them to be undefined,
you have to explicitly set them `undefined`). Also, this strategy
cannot get used in every circumstance, for example if you object
uses other classes as members. In that case you rather write a
`reinit()` or similar function, which resets the state of your object.